DFK-262 - Flyttet måling-testregel logikk til TestregelResource og testregel-api
